+++ src/plooney81/nectar/sql/impl.clj
@@ -26,6 +26,12 @@
     (.getClass jsql-update)))
 
 (defmulti
+  delete->honey
+  "Converts jsql-delete objects to honey-sql based on the class"
+  (fn [_honey jsql-delete]
+    (.getClass jsql-delete)))
+
+(defmulti
   expression
   "Converts a jsql expression to an alternative syntax that will ultimately be used to convert to honeysql"
   (fn [jsql-expr]