+++ src/plooney81/nectar/sql/update.clj
@@ -13,6 +13,11 @@
                         (impl/expression->honey (first (.getValues update-set)))))
                {})))
 
+(defn- handle-with-items [honey ^Update jsql-update]
+  (if-let [with-items (.getWithItemsList jsql-update)]
+    (helpers/convert-with-list honey with-items)
+    honey))
+
 (defn- handle-table [honey ^Update jsql-update]
   (sql/update honey (jsql/convert-table (.getTable jsql-update))))
 
@@ -34,6 +39,11 @@
     (apply sql/order-by honey order-by-items)
     honey))
 
+(defn- handle-joins [honey ^Update jsql-update]
+  (if-let [joins (.getJoins jsql-update)]
+    (helpers/convert-join-list honey joins)
+    honey))
+
 (defn- handle-limit [honey ^Update jsql-update]
   (if-let [limit (.getLimit jsql-update)]
     (sql/limit honey (jsql/get-limit-value limit))
@@ -41,9 +51,11 @@
 
 (defmethod impl/update->honey Update [honey ^Update jsql-update]
   (-> honey
+      (handle-with-items jsql-update)
       (handle-table jsql-update)
       (handle-set jsql-update)
       (handle-from jsql-update)
+      (handle-joins jsql-update)
       (handle-where jsql-update)
       (handle-order-by jsql-update)
       (handle-limit jsql-update)))